\name{argList-methods}
\docType{methods}
\alias{argList}
\alias{argList-methods}
\title{Methods for function argList in package 'MareyMap'}
\description{
 Methods for function \code{argList} in package \pkg{MareyMap} returns the argument list for object
}
\section{Methods}{
\describe{
  \item{object = "Interpolation"}{returns the basic interpolation parameters (for use in createOrder)}
}}

\keyword{methods}

\seealso{
  \code{\link{Interpolation-class}}
}

\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}

