\name{color-methods}
\docType{methods}
\alias{color}
\alias{color-methods}
\alias{color<-}
\alias{color<--methods}
\title{Accessors for slot color in package 'MareyMap'}
\description{
 Methods for function \code{color} in package \pkg{MareyMap}
}
\section{Methods}{
\describe{
  \item{object = "Interpolation"}{Returns or sets the value of the slot color in the \code{\link{interpolation}} object}
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}

