\name{createOrder-methods}
\docType{methods}
\alias{createOrder}
\alias{createOrder-methods}
\title{createOrder}
\description{
	Returns a character containing the code to recreate the object as it is at the moment of the call (e.g. "new(objCLass, arg1=...)")
}
\section{Methods}{
\describe{
  \item{object = "MMSlidingWindow"}{createOrder for method MMSlidingWindow}
  \item{object = "MMLoess"}{createOrder for method MMLoess}
  \item{object = "MMSpline3"}{createOrder for method MMSpline3}
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}
