\name{geneticDistances-methods}
\docType{methods}
\alias{geneticDistances}
\alias{geneticDistances-methods}
\alias{geneticDistances<-}
\alias{geneticDistances<--methods}
\title{Accessors for slot geneticDistances}
\description{
  Methods for function \code{geneticDistances} in package \pkg{MareyMap}
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{returns or set the value of the slot geneticDistances in MareyMap}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}
