\name{getInterList-methods}
\docType{methods}
\alias{getInterList}
\alias{getInterList-methods}
\alias{getInterList,missing-method}
\title{returns a list of interpolation methods.}
\description{
  returns the list of the interpolation methods present in the interpolation register.
}
\section{Methods}{
\describe{
  \item{dummy = "missing"}{default method.}
}}
\keyword{methods}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}