\name{interpolation-methods}
\docType{methods}
\alias{interpolation}
\alias{interpolation-methods}
\alias{interpolation<-}
\alias{interpolation<--methods}
\title{set or returns an interpolation.} 
\description{
  Methods for function \code{interpolation} in package \pkg{MareyMap}
}
\section{Methods}{
\describe{
\item{object = "MareyMap", inter\_name = "character"}{set or returns the interpolation \code{inter\_name} in object \code{\link{MareyMap-class}}}
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}