\name{InterpolationParam-methods}
\docType{methods}
\alias{InterpolationParam-methods}
\alias{InterpolationParam,missing-method}
\title{Methods for Function InterpolationParam in Package 'MareyMap'}
\description{
 constructor for class \code{\link{InterpolationParam-class}} in package \pkg{MareyMap}
}
\section{Methods}{
\describe{
  \item{dummy = "missing"}{default constructor.}
}}
\keyword{methods}
\seealso{
  \code{link{InterpolationParam-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}