\name{interpolations-methods}
\docType{methods}
\alias{interpolations}
\alias{interpolations-methods}
\alias{interpolations<-}
\alias{interpolations<--methods}
\title{accessors for slot interpolations}
\description{
 set or get the value of the slot \code{interpolation} in \code{object}
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{ set or get slot \code{interpolations} in \code{\link{MareyMap-class}}}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}