\name{MapCollection-methods}
\docType{methods}
\alias{MapCollection-methods}
\title{Constructor for class MapCollection}
\description{
 Creates an object of class \code{\link{MapCollection}}
}
\section{Methods}{
\describe{
  \item{x = "missing"}{Default constructor.}
}}
\keyword{methods}
\seealso{
  \code{\link{MapCollection-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}