\name{mapName-methods}
\docType{methods}
\alias{mapName}
\alias{mapName-methods}
\alias{mapName<-}
\alias{mapName<--methods}
\title{ accessors for the slot mapName}
\description{
 Set or get the value of the slot \code{mapName} in \code{object}. 
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{  Set or get the value of the slot \code{mapName} in \code{object}.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}