\name{mapNames-methods}
\docType{methods}
\alias{mapNames}
\alias{mapNames-methods}
\title{mapNames}
\description{
 Returns the names of the maps contained in \code{object}.
}
\section{Methods}{
\describe{
  \item{object = "MapSet"}{Returns the names of the maps contained in \code{MapSet}}
}}
\keyword{methods}
\seealso{
  \code{\link{MapSet}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}
