\name{MareyMap-methods}
\docType{methods}
\alias{MareyMap-methods}
\alias{MareyMap,missing,missing,missing,missing-method}
\title{Constructor for object of the class MareyMap.}
\description{
 Creates an object of the class \code{\link{MareyMap-class}}.
}
\section{Methods}{
\describe{
  \item{data\_table = "missing", column\_names = "missing", set\_name = "missing", map\_name = "missing"}{Default constructor}
  \item{data\_table = "data.frame"}{fill the class using the provided data.frame}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}
