\name{markerNames-methods}
\docType{methods}
\alias{markerNames}
\alias{markerNames-methods}
\alias{markerNames<-}
\alias{markerNames<--methods}
\title{Accessors for the class markerNames.}
\description{
 Get or Set slot \code{markerNames} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "MareyMap"}{Get or Set slot \code{markerNames} in \code{\link{MareyMap}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}