\name{markerValidity-methods}
\docType{methods}
\alias{markerValidity}
\alias{markerValidity-methods}
\alias{markerValidity<-}
\alias{markerValidity<--methods}
\title{accessors for slot markerValidity}
\description{
 Get or set the value of the slot \code{MarkerValidity} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "MareyMap"}{Get or set the value of the slot \code{MarkerValidity} in \code{\link{MareyMap}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}