\name{name-methods}
\docType{methods}
\alias{name}
\alias{name-methods}
\alias{name<-}
\alias{name<--methods}
\title{Accessor for slot name.}
\description{
 Get or set the value of the slot \code{name} in \code{object}
}
\section{Methods}{
\describe{
\item{object = "Interpolation"}{Get or set the value of the slot \code{name} in \code{Interpolation}.}
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}