\name{paramDefault-methods}
\docType{methods}
\alias{paramDefault}
\alias{paramDefault-methods}
\alias{paramDefault<-}
\alias{paramDefault<--methods}
\title{Accessors for slot paramDefault}
\description{
 Get or set the value of the slot \code{paramDefault} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{Get or set the value of the slot \code{paramDefault} in \code{\link{MareyMap-class}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}