\name{paramDesc-methods}
\docType{methods}
\alias{paramDesc}
\alias{paramDesc-methods}
\alias{paramDesc<-}
\alias{paramDesc<--methods}
\title{Accessors for the slot paramDesc}
\description{
 Accessors for the slot \code{paramDesc} in \code{object}
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{Accessors for the slot \code{paramDesc} in \code{\link{InterpolationParam}}}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}