\name{paramFun-methods}
\docType{methods}
\alias{paramFun}
\alias{paramFun-methods}
\alias{paramFun<-}
\alias{paramFun<--methods}
\title{ accessors for the slot paramFun}
\description{
 Get or set the value of the slot \code{paramFun} in \code{object}
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{Get or set the value of the slot \code{paramFun} in \code{\link{InterpolationParam-class}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}
