
\name{paramMax-methods}
\docType{methods}
\alias{paramMax}
\alias{paramMax-methods}
\alias{paramMax<-}
\alias{paramMax<--methods}
\title{Accessors for the slot paramMax.}
\description{
 Get or set the value of the slot \code{paramMax} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{Get or set the value of the slot \code{paramMax} in \code{\link{InterpolationParam}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}