\name{paramMin-methods}
\docType{methods}
\alias{paramMin}
\alias{paramMin-methods}
\alias{paramMin<-}
\alias{paramMin<--methods}
\title{Accessors for slot paramMin.}
\description{
 Get or set the value of the parameter \code{paramMin} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{Get or set the value of the parameter \code{paramMin} in \code{InterpolationParam}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}
