\name{paramName-methods}
\docType{methods}
\alias{paramName}
\alias{paramName-methods}
\alias{paramName<-}
\alias{paramName<--methods}
\title{Accessors for the slot paramName.}
\description{
 Get or set the value of the slot \code{paramName} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{ Get or set the value of the slot \code{paramName} in \code{\link{InterpolationParam}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}