\name{paramType-methods}
\docType{methods}
\alias{paramType}
\alias{paramType-methods}
\alias{paramType<-}
\alias{paramType<--methods}
\title{Accessors for the slot paramType.}
\description{
 Get or set the value of the slot \code{paramType} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{ Get or set the value of the slot \code{paramType} in \code{\link{InterpolationParam}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}