\name{paramValues-methods}
\docType{methods}
\alias{paramValues}
\alias{paramValues-methods}
\alias{paramValues<-}
\alias{paramValues<--methods}
\title{Accessor for the slot paramValues.}
\description{
 Get or set the value of the slot \code{paramValues} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "InterpolationParam"}{ Get or set the value of the slot \code{paramValues} in \code{\link{InterpolationParam}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{InterpolationParam-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}