\name{persistent-methods}
\docType{methods}
\alias{persistent}
\alias{persistent-methods}
\alias{persistent<-}
\alias{persistent<--methods}
\title{Accessors for the slot persistent.}
\description{
 Get or set the value of the slot \code{persistent} in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "Interpolation"}{ Get or set the value of the slot \code{persistent} in \code{\link{Interpolation}}. }
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}