\name{physicalPositions-methods}
\docType{methods}
\alias{physicalPositions}
\alias{physicalPositions-methods}
\alias{physicalPositions<-}
\alias{physicalPositions<--methods}
\title{Accessors for the slot physicalPositions}
\description{
 Get or set the value of the slot \code{physicalPositions} \code{object}
}
\section{Methods}{
\describe{
\item{object = "MareyMap"}{Get or set the value of the slot \code{physicalPositions} \code{\link{MareyMap-class}}}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}