\name{plotMarkers-methods}
\docType{methods}
\alias{plotMarkers}
\alias{plotMarkers-methods}
\title{plotMarkers}
\description{
	Plots the markers contained in \code{object}.
}
\section{Methods}{
\describe{
\item{object = "MareyMap"}{Plots the markers of the map \code{\link{MareyMap}}}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}