\name{plotModels-methods}
\docType{methods}
\alias{plotModels}
\alias{plotModels-methods}
\title{plotModels }
\description{
 plot models from \code{object}
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{ Calls plotModel on all the interpolation defined on the \code{\link{MareyMap}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}