\name{plotRates-methods}
\docType{methods}
\alias{plotRates}
\alias{plotRates-methods}
\title{ plotRates}
\description{
	Plots the variations of the recombination rate as calculated by all the
	interpolations defined on the object.
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{ Plots the variations of the recombination rate as calculated by all the interpolations defined on the \code{\link{MareyMap}}. }
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}