\name{rates-methods}
\docType{methods}
\alias{rates}
\alias{rates-methods}
\alias{rates<-}
\alias{rates<--methods}
\title{Accessors for the slot rates}
\description{
 Get or set the value of the slot \code{rates} in \code{object}
}
\section{Methods}{
\describe{
  \item{object = "Interpolation"}{Get or set the value of the slot \code{rates} in \code{\link{Interpolation}}.}
}}
\keyword{methods}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}
