\name{registerInterpolationMethod-methods}
\docType{methods}
\alias{registerInterpolationMethod}
\alias{registerInterpolationMethod-methods}
\title{ registerInterpolationMethods}
\alias{registerInterpolationMethod,character,character-method}
\description{
Adds an \code{\link{Interpolation}} method to the interpolation register.
}
\section{Methods}{
\describe{
\item{name = "character", classname = "character"}{Register an interpolation method with a given \code{name} and a contructor function \code{classname}.}
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}