\name{removeMarker-methods}
\docType{methods}
\alias{removeMarker}
\alias{removeMarker-methods}
\title{ removeMarker }
\description{
 Removes a marker from an \code{object} 
}
\section{Methods}{
\describe{
  \item{object = "MareyMap", value = "integer"}{Removes the marker number \code{value} from a \code{\link{MareyMap-class}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}