\name{MapSet-methods}
\docType{methods}
\alias{MapSet-methods}
\alias{MapSet,character-method}
\title{ Constructor for the class MapSet.}
\description{
	Creates an object of the class \code{\link{MapSet}}.
}
\section{Methods}{
\describe{
\item{set\_name = "character"}{ creates a new object of the class \code{\link{MapSet}} with the name set\_name. }
}}
\keyword{methods}

\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}