\name{setName-methods}
\docType{methods}
\alias{setName}
\alias{setName-methods}
\alias{setName<-}
\alias{setName<--methods}
\title{Accessors for the slot setName.}
\description{
 Get or set the value of the slot \code{setName} in \code{object}.
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{Get or set the value of the slot \code{setName} in \code{\link{MareyMap}}.}
  \item{object = "MapSet"}{Get or set the value of the slot \code{setName} in \code{\link{MapSet}}.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap}}
  \code{\link{MapSet}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}