\name{setNames-methods}
\docType{methods}
\alias{setNames}
\alias{setNames-methods}
\title{setNames}
\description{
 returns the names of the set contained in object.
}
\section{Methods}{
\describe{
  \item{object = "MapCollection"}{ Returns the name of the set contained in \code{\link{MapCollection}}. }
}}
\keyword{methods}
\seealso{
  \code{\link{MapCollection}}.
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}