\name{startMareyMapGUI-methods}
\docType{methods}
\alias{startMareyMapGUI}
\alias{startMareyMapGUI-methods}
\alias{startMareyMapGUI,missing-method}
\title{MareyMapGUI launcher}
\description{
 launches the graphical user interface of MareyMapGUI
}
\section{Methods}{
\describe{
  \item{dummy = "missing"}{Default method.}
}}
\keyword{methods}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}