\name{textFile-methods}
\docType{methods}
\alias{textFile}
\alias{textFile-methods}
\title{textFile}
\description{
	Writes object to text file.
}
\section{Methods}{
\describe{
  \item{object = "MapCollection", file = "character"}{Writes a \code{\link{MapCollection}} to text file.}
  \item{object = "MareyMap", file = "character"}{Writes a \code{\link{MareyMap}} to text file.}
  \item{object = "MapSet", file = "character"}{Writes a \code{\link{MapSet}} to text file.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
  \code{\link{MapCollection-class}}
  \code{\link{MapSet-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}