\name{userParam-methods}
\docType{methods}
\alias{userParam}
\alias{userParam-methods}
\title{ userParam} 
\description{
	Returns information about the parameters modifiable by the user in object
}
\section{Methods}{
\describe{
  \item{object = "Interpolation"}{Returns information about the parameters modifiable by the user in \code{\link{Interpolation}}}
  \item{object = "MMSlidingWindow"}{ Returns information about the parameters modifiable by the user in MMSlidingWindow} 
  \item{object = "MMLoess"}{ Returns information about the parameters modifiable by the user in MMLoess}
  \item{object = "MMSpline3"}{ Returns information about the parameters modifiable by the user in  MMSpline3}
}}
\keyword{methods}
\seealso{
  \code{\link{Interpolation-class}}
  \code{\link{InterpolationParam-class}}
}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}