\name{validPositions-methods}
\docType{methods}
\alias{validPositions}
\alias{validPositions-methods}
\title{validPositions}
\description{
 Returns the positions of object which are valid.
}
\section{Methods}{
\describe{
  \item{object = "MareyMap"}{Returns the physical positions of MareyMap which are valid.}
}}
\keyword{methods}
\seealso{
  \code{\link{MareyMap-class}}
}
\author{Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clement Rezvoy}