\name{visible-methods}
\docType{methods}
\alias{visible}
\alias{visible-methods}
\alias{visible<-}
\alias{visible<--methods}
\title{Accessors for the slot visible}
\description{
 Get of set the value of the slot \code{visible} in \code{object}
}
\section{Methods}{
\describe{
  \item{object = "Interpolation"}{Get of set the value of the slot \code{visible} in \code{\link{Interpolation-class}}}
}}
\keyword{methods}
\author{Aurélie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Clément Rezvoy}